fix(render, layout): prevent arithmetic underflow in cell run and overflow redistribution
In the render diff engine, ChangeRun::len() could underflow when x1 < x0
due to unsigned subtraction, producing a wrapped u16 value. Similarly,
the presenter cost model computed total_cells with a bare subtraction
that could panic or wrap. Both now use saturating_sub to safely handle
edge cases where coordinate ordering is unexpected.

In the layout overflow redistribution, the min-chain operated on u16
values which could silently truncate when overflow exceeded u16::MAX.
Casting to u64 before comparison ensures correctness for large overflow
values that arise with many layout segments.

+++ crates/ftui-layout/src/lib.rs
@@ -1187,7 +1187,7 @@ fn redistribute_overflow(floors: &[u16], total: u16) -> Vec<u16> {
                 break;
             }
             if result[idx] > 0 {
-                let amount = result[idx].min(reduce_amount).min(overflow as u16);
+                let amount = (result[idx] as u64).min(reduce_amount as u64).min(overflow) as u16;
                 result[idx] -= amount;
                 overflow -= amount as u64;
                 reduced_any = true;
